refactor(app): simplify function runtime loading
